+++ example/index_generator_example.dart
@@ -1,8 +1,0 @@
-import 'package:index_generator/src/index_generator.dart';
-
-import '../bin/main.dart' as generator;
-
-void main() {
-  isTestingMode = true;
-  generator.main();
-}